Colocate transforms control with Link UI component
